+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ## Unreleased
 When you submit a PR, add your changes here!
+### Added
+- Use `headerLabelFormats` and `subHeaderLabelFormats` to customise the header labels. @Slowyn #68
 
 ## [0.13.0]
 ### Added